requirements.txt
openai
google-generativeai
supabase